% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_autorelevate.R
\name{print.autorelevate_fit}
\alias{print.autorelevate_fit}
\title{Print Summary of an Autorelevated Fit}
\usage{
\method{print}{autorelevate_fit}(x, ...)
}
\arguments{
\item{x}{An autorelevate_fit object.}

\item{...}{Unused additional arguments.}
}
\value{
Invisibly, x.
}
\description{
Print Summary of an Autorelevated Fit
}
\examples{
set.seed(1)
x <- rautorelevate(100, dist = "weibull", p1 = 0.5, p2 = 1.5)
print(fit_autorelevate(x, dist = "weibull", method = "mle"))
}
